+++ apps/eodh-fe/public/assets/i18n/en.json
@@ -186,7 +186,6 @@
         "NO_RESULTS_MESSAGE": "You haven't run any workflows yet. Start by creating a new workflow in the Workflow tab.",
         "VIEW_RESULTS": "View results",
         "HIDE_RESULTS": "Hide results",
-        "DELETE": "Delete",
         "SAVED_ON": "Saved on",
         "SAVED_AT": "at",
         "ID": "ID",
