--- libs/map/feature-action-creator-panel/src/lib/content/history/history-tile/tag.component.tsx
@@ -1,20 +1,0 @@
-import { Text } from '@ukri/shared/design-system';
-import clsx from 'clsx';
-
-export const Tag = ({ status }: { status: 'READY' | 'PROCESSING' | 'FAILED' }) => {
-  const tagStyles = {
-    READY: 'bg-success text-success-contrastText',
-    PROCESSING: 'bg-warning text-bright',
-    FAILED: 'bg-error text-error-contrastText',
-  };
-  return (
-    <div className={clsx('rounded h-5 flex items-center', tagStyles[status])}>
-      <Text
-        content={`MAP.ACTION_CREATOR_PANEL.HISTORY.STATUS.${status}`}
-        fontSize='small'
-        fontWeight='bold'
-        className='mx-1.5 my-[3px] uppercase'
-      />
-    </div>
-  );
-};
