+++ libs/shared/theme/src/lib/ligh-theme.config.ts
@@ -33,11 +33,6 @@ const colors = {
     main: '#FF0000',
     contrastText: '#FFFFFF',
   },
-  errorLight: {
-    DEFAULT: '#F5CFCF',
-    main: '#F5CFCF',
-    contrastText: '#FF0000',
-  },
   warning: {
     DEFAULT: '#FFD15B',
     warning: '#FFD15B',
